requirements.txt
# Lighter vector database option
faiss-cpu==1.7.4
openai==1.3.7
python-dotenv==1.0.0

# File processing
python-frontmatter==1.0.0

# Data handling
numpy==1.24.3
pandas==2.0.3

# Optional: for future web interface
streamlit==1.28.1